# app/src/main/java/io/swagger/dto/model/Cart.java
package io.swagger.dto.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;
import java.util.UUID;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class Cart {

    private UUID cartId;

    private Customer customer;

    private List<Item> cartItems;

}