# app/src/main/java/io/swagger/dto/model/Customer.java
package io.swagger.dto.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class Customer {

    private String email;

    private String name;

    private String surname;
}