# app/src/main/java/io/swagger/dto/request/CustomerRequest.java
package io.swagger.dto.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class CustomerRequest {

    private String email;

    private String name;

    private String surname;


}